Add files via upload
Acknowledgements
Acknowledgement. Authors must include individuals and organizations that have made substantive contributions to the research.

Acknowledgement section not found: If appropriate, please add this section and include individuals and organizations that have made substantive contributions to the research or the manuscript. An exception is where funding was provided, which should be included in Funding Sources. Please refer to the Guidelines issued by the ICMJE to determine non-author contributors that should be included in the Acknowledgement section.

Ethics
Studies involving human subjects

Informed consent not addressed: Studies involving human subjects (including research on identifiable human material and data) must have been performed with the appropriate participants' informed consent in compliance with the Helsinki Declaration.

Have you included your trial registration ID? Generally this is compulsory if your manuscript is a clinical trial, and it should go at the end of your abstract. You should also include the registry name and date of registration. If your trial was registered retrospectively you should state that too.

Ethics section header not found: All submitted manuscripts must contain a Statement of Ethics. Published research must comply with internationally accepted standards for research practice and reporting.Manuscripts may be rejected if the editors believe that the research has not been carried out within an appropriate ethical framework, and concerns raised after publication may lead to a correction, retraction, or expression of concern in line with COPE guidelines.

Ethical declaration found: Ensure your declaration follows internationally recognized guidelines as detailed on the Equator network website.

Ethical Review Board named: You have specified the name of the ethics committee or other relevant authority who approved the study protocol. Please also ensure that you provide the reference number.

Conflict of Interest Statement
Authors are required to disclose any possible conflicts of interest in the manuscript file.

Conflict of Interest header found. Please remember to disclose any possible conflicts of interest. All forms of support and financial involvement (e.g. employment, consultancies, honoraria, stock ownership and options, expert testimony, grants or patents received or pending, royalties) which took place in the previous three years should be listed, regardless of their potential relevance to the paper. Also the nonfinancial relationships (personal, political, or professional) that may potentially influence the writing of the manuscript should be declared. If there is no conflict of interest, please state: "The authors have no conflicts of interest to declare."

Funding
Authors must give full details about the funding of any research relevant to their study.

Funding section found: Remember to include details about the funding of any research relevant to the study, including sponsor names and explanations of the roles of these sources.

Author contributions
In the Author Contributions section, a short statement detailing the contributions of each person named as an author should be included. Contributors to the paper who do not fulfill the lICMJE Criteria for Authorship should be credited in the Acknowledgement section - Contributors to the paper are to fulfill the ICMJE Criteria for Authorship - See documentation

Author Contribution section not found: A short statement detailing the contributions of each person named as an author should be included.
수정